feat: execute_one_time_command()
